requirements.txt
adblockparser==0.7
appdirs==1.4.3
asn1crypto==0.22.0
attrs==17.1.0
Automat==0.6.0
certifi==2017.4.17
cffi==1.10.0
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
decorator==4.0.11
Django==1.11.1
django-dynamic-scraper==0.12.0
django-reversion==2.0.8
djangorestframework==3.6.3
funcparserlib==0.3.6
future==0.16.0
idna==2.5
incremental==16.10.1
jsonpath-rw==1.4.0
lxml==3.7.3
olefile==0.44
packaging==16.8
parsel==1.2.0
Pillow==4.1.1
ply==3.10
psutil==5.2.2
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pyparsing==2.2.0
pypiwin32==219
python-telegram-bot==6.0.1
pytz==2017.2
qt5reactor==0.4
queuelib==1.4.2
requests==2.14.2
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
scrapy-splash==0.7.2
service-identity==16.0.0
six==1.10.0
splash==2.3.2
toml==0.9.2
Twisted==17.1.0
urllib3==1.21.1
w3lib==1.17.0
xvfbwrapper==0.2.9
zope.interface==4.4.1
